chore: rm unused package & refactor methods
