+++ packages/amplify-codegen-e2e-tests/package.json
@@ -31,7 +31,6 @@
     "aws-sdk": "^2.1413.0",
     "dotenv": "^8.2.0",
     "esm": "^3.2.25",
-    "execa": "^8.0.1",
     "fs-extra": "^8.1.0",
     "graphql-tag": "^2.10.1",
     "js-yaml": "^4.0.0",
@@ -43,6 +42,7 @@
   },
   "devDependencies": {
     "@types/jest": "^27.0.0",
+    "@aws-amplify/backend": "^1.0.0",
     "graphql": "^15.8.0",
     "ts-node": "^8.10.1"
   },
