+++ packages/amplify-codegen-e2e-tests/src/__tests__/graphql-generator-gen2.test.ts
@@ -1,5 +1,6 @@
+import * as path from 'path';
 import { createNewProjectDir, deleteProjectDir } from "@aws-amplify/amplify-codegen-e2e-core";
-import { deleteSandbox, initGen2Project, sandboxDeploy } from "../commands";
+import { deleteSandbox, initGen2Project, sandboxDeploy } from "../gen2-codegen-tests-base/commands";
 
 describe('GraphQL generator for Gen2 e2e tests', () => {
   let projRoot: string;
@@ -16,7 +17,8 @@ describe('GraphQL generator for Gen2 e2e tests', () => {
   });
 
   it('demo', async () => {
-    const name = await initGen2Project(projRoot);
+    const template = path.resolve(path.join(__dirname, 'backends', 'graphql-generator-gen2'))
+    await initGen2Project(projRoot, template);
     await sandboxDeploy(projRoot);
   })
 });
