+++ packages/amplify-codegen-e2e-tests/src/gen2-codegen-tests-base/commands.ts
@@ -1,13 +1,20 @@
 import * as path from 'path';
 import { copySync, moveSync, readFileSync, writeFileSync } from 'fs-extra';
-import { getScriptRunnerPath, nspawn as spawn, npmInstall, getCommandPath } from '@aws-amplify/amplify-codegen-e2e-core';
-import { spawnSync, execSync } from 'child_process';
+import { getScriptRunnerPath, nspawn as spawn, getCommandPath } from '@aws-amplify/amplify-codegen-e2e-core';
+import { spawnSync } from 'child_process';
 
 /**
  * Retrieve the path to the `npx` executable for interacting with the amplify gen2 cli.
  * @returns the local `npx` executable path.
  */
 const getNpxPath = (): string => (process.platform === 'win32' ? getScriptRunnerPath().replace('node.exe', 'npx.cmd') : 'npx');
+/**
+ * Retrieve the path to the `ampx` executable for interacting with the amplify gen2 cli.
+ * @param cwd current working directory
+ * @returns the local `ampx` executable path
+ */
+const getAmpxPath = (cwd: string): string => 
+  spawnSync(getNpxPath(), ['which', 'ampx'], { cwd, env: process.env, stdio: 'pipe' }).stdout.toString().trim();
 
 const codegenPackagesInGen2 = [
   '@aws-amplify/graphql-generator',
@@ -20,17 +27,16 @@ const apiPackagesInGen2 = [
 type CodegenPackage = 'GraphqlGenerator' | 'TypeGen';
 
 const codegenPackageDirectoryMap: Record<CodegenPackage, string> = {
-  GraphqlGenerator: path.join(__dirname, '..', '..', 'graphql-generator'),
-  TypeGen: path.join(__dirname, '..', '..', 'graphql-types-generator'),
+  GraphqlGenerator: path.join(__dirname, '..', '..', '..', 'graphql-generator'),
+  TypeGen: path.join(__dirname, '..', '..', '..', 'graphql-types-generator'),
 }
 
 /**
- * Copy the backend snapshot into the generated app location.
+ * Copy the backend data snapshot into the generated app location.
  */
 const copyTemplateDirectory = (projectPath: string, templatePath: string): void => {
-  const binDir = path.join(projectPath, 'bin');
-  copySync(templatePath, binDir, { overwrite: true });
-  moveSync(path.join(binDir, 'app.ts'), path.join(binDir, `${path.basename(projectPath)}.ts`), { overwrite: true });
+  const amplifyDataDir = path.join(projectPath, 'amplify', 'data');
+  copySync(templatePath, amplifyDataDir, { overwrite: true });
 };
 
 /**
@@ -40,37 +46,26 @@ const copyTemplateDirectory = (projectPath: string, templatePath: string): void 
  * @param props additional properties to configure the test app setup.
  * @returns a promise which resolves to the stack name
  */
-export const initGen2Project = async (cwd: string, props: Gen2DeployProps = {}): Promise<string> => {
+export const initGen2Project = async (cwd: string, templatePath: string, props: Gen2DeployProps = {}): Promise<string> => {
   const commandOptions = {
     cwd,
     stripColors: true,
   };
   const npmPath = getCommandPath('npm')
-  // const commandOptions = {
-  //   cwd,
-  //   env: process.env,
-  //   stdio: 'inherit'
-  // }
-  // spawnSync('npm', ['create', 'amplify@latest', '-y'], commandOptions as any);
   await spawn(npmPath, ['create', 'amplify@latest', '-y'], commandOptions).runAsync();
-
-  const ampx = spawnSync(getNpxPath(), ['which', 'ampx'], { cwd, env: process.env, stdio: 'pipe' }).stdout.toString().trim()
-  console.log(ampx)
 
   overrideWithLocalCodegenPackages(cwd);
 
-  // spawnSync('npm', ['install'], commandOptions as any);
   await spawn(npmPath, ['install'], commandOptions).runAsync();
 
   // Get root level packages info
-  // spawnSync('npm', ['list'], commandOptions as any)
   await spawn(npmPath, ['list'], commandOptions).runAsync();
   // Get codegen packages info
-  // spawnSync('npm', ['list', ...codegenPackagesInGen2], commandOptions as any)
   await spawn(npmPath, ['list', ...codegenPackagesInGen2], commandOptions).runAsync();
   // Get API packages info
-  // spawnSync('npm', ['list', ...apiPackagesInGen2], commandOptions as any)
   await spawn(npmPath, ['list', ...apiPackagesInGen2], commandOptions).runAsync();
+
+  copyTemplateDirectory(cwd, templatePath);
 
   return JSON.parse(readFileSync(path.join(cwd, 'package.json'), 'utf8')).name.replace(/_/g, '-');
 };
@@ -95,14 +90,17 @@ export const sandboxDeploy = async (cwd: string, props: Gen2DeployProps = {}) =>
     stripColors: true,
     noOutputTimeout,
   };
+
   // Run sandbox deployment
-  const ampx = spawnSync(getNpxPath(), ['which', 'ampx'], { cwd, env: process.env, stdio: 'pipe' }).stdout.toString().trim()
-  // const ampx = execSync('npx which ampx')
-  //   .toString()
-  //   .trim();;
-  console.log(ampx)
+  
+  /**
+   * For sandbox deployment, the nested ampx binary is retrieved instead of using npx
+   * On windows, the Ctrl-C signal is not returned correctly from npx binary, whose code is 1 and will fail nexpect check
+   * Therefore, the ampx binary is used for sandbox deployment instead of npx
+   */
+  const ampxCli = getAmpxPath(cwd)
   await
-    spawn(ampx, ['sandbox'], commandOptions)
+    spawn(ampxCli, ['sandbox'], commandOptions)
       .wait('Watching for file changes...')
       .sendCtrlC()
       .wait('Would you like to delete all the resources in your sandbox environment')
