+++ packages/amplify-codegen-e2e-tests/src/gen2-codegen-tests-base/index.ts
@@ -1,0 +1,1 @@
+export * from './commands';
